requirements.txt
squareup < 42.0.0
pandas
dotenv
streamlit